Remove index on vulnerabilities
We should have an index on vulnerabilities because we query this table
using name, description, fixedby and link. Since they are so big, it
needs to be a functional index computing a hash (e.g. md5) of the column
and the related queries to this table need to be modified to use it.

We need to assert the need for that work, so it will done in a
subsequent effort as it is blocking the rest of the indexes that should
have been created as part of #53

Since schema modifications are transactional in PostgreSQL, we don't
need a new migration as the migration from #53 was cleanly rolled back
in staging environment, hence this change just removes the index
creation and deletion from the migration file of #53.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>