specify that value and target are not nullable
The database migration to change `value` and `target` to type float was
created manually, and now every time we run `flask db migrate` it tries
to create a new migration to change those values to nullable.

This PR adds `nullable=False` to the data types so it matches the
current state of the database, therefore this avoids migrations.