APPSRE-11100: Fixup dashdotdb uv
The previous commits (865b64ba and cab5fc9d) didn't quite get dashdotdb working in stage because of OCP runtime restrictions that were not evident in my local development and testing. This commit, on the other hand, should be expected to work because the `prod` image will not rely on `uv` to run flask and gunicorn.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>